// frontend/src/generated/graphql.ts
export enum Priority {
  HIGH = 'HIGH',
  MEDIUM = 'MEDIUM',
  LOW = 'LOW',
}

export enum TaskStatus {
  TODO = 'TODO',
  IN_PROGRESS = 'IN_PROGRESS',
  DONE = 'DONE',
}

export interface Subject {
  __typename?: 'SubjectModel';
  id: string;
  name: string;
  color: string;
  order: number;
  units: Unit[];
}

export interface Unit {
  __typename?: 'UnitModel';
  id: string;
  name: string;
  order: number;
  subjectId: string;
}
// ...
export interface TaskFilterInput {
  subjectId?: string | null;
  unitId?: string | null;
  priority?: Priority | null;
  status?: TaskStatus | null;
}